requirements.txt
playwright
python-slugify
